Speed up pkg.grep by utilizating mmap.
The following changes speed up kernel-source package by 23s.
