+++ rpmlint/configdefaults.toml
@@ -3,8 +3,7 @@ Checks = [
     "AlternativesCheck",
     "AppDataCheck",
     "BinariesCheck",
-    "BuildDateCheck",
-    'BuildRootCheck',
+    'BuildRootAndDateCheck',
     "ConfigFilesCheck",
     "DBusPolicyCheck",
     'DuplicatesCheck',
