+++ test/test_lint.py
@@ -31,8 +31,7 @@ basic_tests = [
     'AlternativesCheck',
     'AppDataCheck',
     'BinariesCheck',
-    'BuildDateCheck',
-    'BuildRootCheck',
+    'BuildRootAndDateCheck',
     'ConfigFilesCheck',
     'DBusPolicyCheck',
     'DuplicatesCheck',
